Fix transitionSpeed getting stuck on 0ms
Because the configured transition speed was doubling as a state
variable, certain event sequences could cause the configured value to be
forgotten completely, resulting in a permanent 0ms transitionSpeed.

This revision separates the state variable from the configuration
parameter.

The nondeterminism is caused by use of timers for sequencing. It might
be possible to improve this eventually but the fix presented here is
much more straightforward.